Added printf format string argument checking under gcc to std.h
Corrected or commented existing errors.

NB: gx_color_index maybe 4 or 8 giving interesting results to printf
that choose to do:  "color is %lx", color

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6454 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/gsfcmap1.c
@@ -176,7 +176,7 @@ code_map_decode_next_multidim_regime(const gx_code_map_t * pcmap,
         dlprintf("[J]CMDNmr() is called: str=(");
         debug_print_string_hex(str, ssize);
         dlprintf3(") @ 0x%lx ssize=%d, %d ranges to check\n",
-                       str, ssize, pcmap->num_lookup);
+		  (ulong)str, ssize, pcmap->num_lookup);
     }
  
     for (i = pcmap->num_lookup - 1; i >= 0; --i) {
